[AW-1606] AW - Add username / env selection for MCP Client
