+++ client/env.example
@@ -5,17 +5,12 @@
 
 # MCP Server Configuration
 MCP_SERVER_URL=https://mcp.insightfinder.com
-INSIGHTFINDER_API_URL=https://stg.insightfinder.com
-INSIGHTFINDER_USER_NAME=
-INSIGHTFINDER_LICENSE_KEY=your_insightfinder_license_key
+HTTP_AUTH_METHOD=api_key
+HTTP_API_KEY=your_mcp_server_api_key
 
 # SSL Configuration
 VERIFY_SSL=true
 SSL_CERT_PATH=
-
-# MCP Server Authentication
-HTTP_AUTH_METHOD=api_key
-HTTP_API_KEY=your_mcp_server_api_key
 
 # =============================================================================
 # LLM API Keys - Add your API keys for the models you want to use
@@ -50,7 +45,7 @@ DEEPSEEK_BASE_URL=https://api.deepseek.com
 DEFAULT_LLM=chatgpt
 
 # Chat History
-TRIM_HISTORY=0
+TRIM_HISTORY=6
 
 # SSE Streaming
 SSE_SHOW_PROGRESS=false
