+++ client/requirements.txt
@@ -7,6 +7,7 @@ langgraph>=0.1.0
 httpx>=0.25.0
 pydantic>=2.0.0
 python-dotenv>=1.0.0
+PyYAML>=6.0.0
 
 # OpenTelemetry (optional, for tracing)
 opentelemetry-api>=1.20.0
